requirements.txt
imagecodecs==2024.9.22
joblib==1.4.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
PyWavelets==1.7.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
wheel==0.45.0
rasterio==1.2